ensureu-ai/requirements.txt
# EnsureU AI Service Dependencies

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LLM Clients
anthropic>=0.36.0
openai>=1.50.0
httpx>=0.27.0

# Database
motor>=3.6.0
pymongo>=4.9.0

# Authentication
PyJWT>=2.9.0

# Vector Store & Embeddings
numpy>=1.26.0
# sentence-transformers>=3.0.0  # Optional: for local embeddings

# Rate Limiting
slowapi>=0.1.9

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
ruff>=0.6.0

# Optional: Redis for caching
# redis>=5.0.0
